Refactor strange naming
